requirements.txt
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
pytest>=7.0.0
lxml>=5.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0